Recreated text formatter, styling of markup
Finally the markup formatter is working as should with the parserObject
data.
There are still some problems to figure out like looking at lines
instead of words to figure out a general line-height for the whole line
of text. Another great thing not to miss out, would be justified text,
but since we are looking at word by word instead of the whole line, its
difficult to predict where there could be made wordspacing and where
not, also in reason of varying font sizes and such.

Anyhow, parsing and formatting seams to work fine for now.