Adding indexes/data points for triggerWords
Now indexes, stored in parserObject.dataPoints are used to identify the
approximate start of a trigger inside a text, there is some tolerance
added which comes into play after strange behaviour of counting of
trimmed data, which was working at some point, but broke on other cases.

This is somehow a workaround but allows the formatter to only start
formating a triggerword if the index in the sentence/activeline of this
word is approperiate.