Tested working with IE10, Firefox and Chrome
Changed some code to be compatible with IE10, unfortunately I cannot
test with IE9 - also Chrome has been tested to be working.

Updated the readme to included the new versions and the formatting guide
to be equal with the lastest version.

Removed canvasLister.js which has been renamed to
canvasLister_improvedParse.js - with the more advanced parser.

Added canvasLister_phantom.js which in addition to the advanced parser
also included the option to set the textAlignment parameter to either
left or justified.

Added version parameter list overview for all versions in calling html
as well as calling syntax to test these.