Fixing parsing, changed pagination, adding resize feature..
This commit fixes some bugs from the tag source parser and also fixes an
issue that the background color, when the canvas is displayed as an
image, is not filled entirely with the specified background color.

The pagination version now creates a canvas which can be activated by
clicking it, then pressing left or right arrow on the keyboard will
allow to navigate through the pages of that canvas.
In addition, this version includes the option to resize the activated
canvas, by pressing plus or minus on the numpad block, if activated to
enlarge or shrink the reading area.

The source markup text has been updated with the new feature information
and the html reduced to only load the most recent versions.