Introducing alt and title text for screenreaders, adding more code comments
The title and alt text attributes are now populated with the data of the
canvas. If there is a title provided, we only set the alt text, in case
there aint any title or alt text - we make use of the first lines of the
source data for the title, the remaining text is set for the alt text,
and so forth.
This might help screenreaders a little to work with the canvas by
providing at least a title and alt text with the text data of the
canvas.

Added some more comments or changed to better explain the code.