Adding image layouting functionality
Images now behave like images in html, meaning that they can be layouted
using the align keyvalue.
phantom_image.js now allows centered images, which occupy a whole
virtual row. Left or right alignment which float like in regular html on
the same row, given the width of the canvas is width is not exceeded. If
so, a overflow happens but the image is placed on the side which is set
with the align key.
The source3.txt and html is changed to show this new behaviour.