Added combined image and pagination version, code cleanup
Phantom_v2.js does now combine the image display with the pagination,
but the feedback of the resize is yet little slow due to image drawing
onto the pages after markup processing.

Cleaned up some variables which are not longer required.
Updated the formatting guide and changed some wording and comments on
feature/functionality.
Updated readme to be more precise.
Added a new source5.txt which has the image test flow cases implemented
and is now be used by the v2 version.