Work on ui responsiveness, img loading, adding imgs
Added several new test images to feed the image loading and layouting.
Added inLayoutProcess and escapeLayoutProcess to make the key handling
more direct and added a check for the image loading if we are in another
layouting process due to resizing of the canvas.
Added "misspelled source?" notification in case the Ajax request to load
the markup data fails to the phantom_image.js version.
Changed markup test-sources to use new images.