Changed image loading, error message clearing
The image loading relies now, besides the width and height of an image,
on its complete property so that the next image only becomes tested if
the image is really loaded completely. Missing this images in Chrome
didnt become fully drawn.

Also the error message clearing has been changed to be only done after
the Ajax request is completed, IE10 and Chrome did clearing during the
content had been drawn.

Added default fillStyle color for the error message.

Changed translate to be set outside the markup processing so the canvas
content is positioned correct.