Clearing img tags from regular text and markup
Image tags are now cleared from the regular text markup, before - the
images have caused errors due its mixing with the other markup.
For now images are only valid if they are not combined with text in one
line, but there can be a chain of img tags on the same line.